reorganize filter helper
